requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.1.1
Django==2.2.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-machina==1.0.1
django-mptt==0.10.0
django-rest-framework==0.1.0
django-stubs==0.12.1
django-tempus-dominus==5.1.2.4
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
idna==2.8
idna-ssl==1.1.0
Markdown==3.1
markdown2==2.3.7
multidict==4.5.2
mypy==0.701
mypy-extensions==0.4.1
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.2
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
typed-ast==1.3.5
typing-extensions==3.7.2
urllib3==1.24.2
websockets==6.0
Whoosh==2.7.4
yarl==1.3.0
